refactor: Move base nodes to node/ folder
