+++ ml_pipeline_engine/visualization/sample.py
@@ -1,7 +1,7 @@
-from ml_pipeline_engine.base_nodes.processors import ProcessorBase
 from ml_pipeline_engine.dag_builders.annotation import build_dag
 from ml_pipeline_engine.dag_builders.annotation.marks import Input
 from ml_pipeline_engine.dag_builders.annotation.marks import SwitchCase
+from ml_pipeline_engine.node.base_nodes import ProcessorBase
 
 
 class Ident(ProcessorBase):
