+++ tests/dag/test_dag_single.py
@@ -1,8 +1,8 @@
 import typing as t
 
-from ml_pipeline_engine.base_nodes.processors import ProcessorBase
 from ml_pipeline_engine.context.dag import DAGPipelineContext
 from ml_pipeline_engine.dag_builders.annotation import build_dag_single
+from ml_pipeline_engine.node.base_nodes import ProcessorBase
 
 
 class DoubleNumber(ProcessorBase):
