+++ tests/tags/test_tags.py
@@ -2,9 +2,9 @@ import typing as t
 
 import pytest_mock
 
-from ml_pipeline_engine.base_nodes.processors import ProcessorBase
 from ml_pipeline_engine.context.dag import DAGPipelineContext
 from ml_pipeline_engine.dag_builders.annotation.marks import Input
+from ml_pipeline_engine.node.base_nodes import ProcessorBase
 from ml_pipeline_engine.node.enums import NodeTag
 from ml_pipeline_engine.parallelism import processes
 from ml_pipeline_engine.parallelism import threads
